backend/requirements.txt
django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.5
Pillow==10.2.0
scikit-learn==1.3.2
nltk==3.8.1
joblib==1.3.2
numpy==1.26.2
pandas==1.5.3
drf-yasg==1.21.7
python-dotenv==1.0.0
